requirements.txt
hydra-core~=1.0.4
pytorch-lightning
torch~=1.7.1
omegaconf~=2.0.5
torchvision~=0.8.2